requirements.txt
# Core dependencies for CV CLI tool
click==8.1.7
python-docx==1.1.0
python-pptx==0.6.23
google-generativeai>=<version>
